build: Make codenarc script work under Windows
Depending on the user's config (esp. core.autocrlf), line endings may be
changed in source files. In most cases, this isn't an issue, but in
Shell script file, the carriage return character (CR) can cause issues
in interpretation. This can be detected when trying to launch the
docker-compose script and having the "‘sh\r’: No such file or directory"
error message: the "\r" is the carriage return being interpreted as part
of the command.

This commit fixes this by specifying that .sh files should always be
checked out with LF endings, not CRLF. .bat files are additionally set
with CRLF endings.

Additionally:

- The tool_prepare-codenarc has been renamed to
  tool_prepare-condenarc.sh for consistency and so that the
  .gitattributes rule would apply to it without declaring it
  specifically. The documentation and build files have been updated
  accordingly.

- The .gitignore has been adapted to allow .gitattributes and some
  comments have been added to explain the rules for non-experts gitters
  (especially since ! rules have been used).